minifix par

+++ internal/booking/model.go
@@ -6,12 +6,12 @@ import (
 
 type Booking struct {
 	reference   string
-	totalAmount int
+	totalAmount int64
 	bookedAt    time.Time
 }
 
 // New validates the reference, totalAmount and bookedAt and returns a Booking.
-func New(reference string, totalAmount int, bookedAt time.Time) (Booking, error) {
+func New(reference string, totalAmount int64, bookedAt time.Time) (Booking, error) {
 
 	if len(reference) != 6 {
 		return Booking{}, ErrInvalidReference
@@ -39,6 +39,6 @@ func (b Booking) BookedAt() time.Time {
 	return b.bookedAt
 }
 
-func (b Booking) TotalAmount() int {
+func (b Booking) TotalAmount() int64 {
 	return b.totalAmount
 }
